move the argument generation in CMDCommand
