+++ src/aaz_dev/command/model/configuration/_command.py
@@ -1,3 +1,5 @@
+import logging
+
 from schematics.models import Model
 from schematics.types import ModelType, ListType, PolyModelType
 
@@ -7,7 +9,10 @@ from ._fields import CMDDescriptionField, CMDVersionField, CMDCommandNameField
 from ._operation import CMDOperation
 from ._output import CMDOutput
 from ._resource import CMDResource
+from ._utils import CMDArgBuildPrefix, CMDDiffLevelEnum
 from utils import exceptions
+
+logger = logging.getLogger('backend')
 
 
 class CMDCommand(Model):
@@ -30,6 +35,60 @@ class CMDCommand(Model):
 
     class Options:
         serialize_when_none = False
+
+    def generate_args(self):
+        arguments = {}
+        for op in self.operations:
+            for arg in op.generate_args():
+                if arg.var not in arguments:
+                    arguments[arg.var] = arg
+
+        # check argument with duplicated option names
+        dropped_args = set()
+        used_args = set()
+        for arg in arguments.values():
+            used_args.add(arg.var)
+            if arg.var in dropped_args or not arg.options:
+                continue
+            r_arg = None
+            for v in arguments.values():
+                if v.var in used_args or v.var in dropped_args or arg.var == v.var or not v.options:
+                    continue
+                if not set(arg.options).isdisjoint(v.options):
+                    r_arg = v
+                    break
+            if r_arg:
+                if self._can_replace_argument(r_arg, arg):
+                    arg.ref_schema.arg = r_arg.var
+                    dropped_args.add(arg.var)
+                elif self._can_replace_argument(arg, r_arg):
+                    r_arg.ref_schema.arg = arg.var
+                    dropped_args.add(r_arg.var)
+                else:
+                    # let developer handle duplicated options
+                    logger.warning(
+                        f"Duplicated Option Value: {set(arg.options).intersection(r_arg.options)} : "
+                        f"{arg.var} with {r_arg.var} : {self.operations[-1].operation_id}"
+                    )
+
+        arguments = [arg for var, arg in arguments.items() if var not in dropped_args]
+
+        arg_groups = {}
+        for arg in arguments:
+            group_name = arg.group or ""
+            if group_name not in arg_groups:
+                arg_groups[group_name] = {}
+            if arg.var not in arg_groups[group_name]:
+                arg_groups[group_name][arg.var] = arg
+
+        groups = []
+        for group_name, args in arg_groups.items():
+            group = CMDArgGroup()
+            group.name = group_name
+            group.args = [arg for arg in args.values()]
+            groups.append(group)
+
+        self.arg_groups = groups or None
 
     def reformat(self, **kwargs):
         self.resources = sorted(self.resources, key=lambda r: r.id)
@@ -82,3 +141,36 @@ class CMDCommand(Model):
                 "details": err.payload['details']
             }
             raise err
+
+    @staticmethod
+    def _can_replace_argument(arg, old_arg):
+        arg_prefix = arg.var.split('.')[0]
+        old_prefix = old_arg.var.split('.')[0]
+        if old_prefix in (CMDArgBuildPrefix.Query, CMDArgBuildPrefix.Header, CMDArgBuildPrefix.Path):
+            # replace argument should only be in body
+            return False
+        if arg_prefix in (CMDArgBuildPrefix.Query, CMDArgBuildPrefix.Header):
+            # only support path argument to replace
+            return False
+
+        elif arg_prefix == CMDArgBuildPrefix.Path:
+            # path argument
+            arg_schema_required = arg.ref_schema.required
+            arg_schema_name = arg.ref_schema.name
+            try:
+                arg.ref_schema.required = old_arg.ref_schema.required
+                if old_arg.ref_schema.name == "name" and "name" in arg.options:
+                    arg.ref_schema.name = "name"
+                diff = arg.ref_schema.diff(old_arg.ref_schema, level=CMDDiffLevelEnum.Structure)
+                if diff:
+                    return False
+                return True
+            finally:
+                arg.ref_schema.name = arg_schema_name
+                arg.ref_schema.required = arg_schema_required
+        else:
+            # body argument
+            diff = arg.ref_schema.diff(old_arg.ref_schema, level=CMDDiffLevelEnum.Structure)
+            if diff:
+                return False
+            return True
