+++ src/aaz_dev/swagger/controller/command_generator.py
@@ -18,9 +18,6 @@ logger = logging.getLogger('backend')
 
 
 class BuildInVariants:
-    Query = "$Query"
-    Header = "$Header"
-    Path = "$Path"
 
     Instance = "$Instance"
 
@@ -165,8 +162,7 @@ class CommandGenerator:
         command.description = op.description
         command.operations = [op]
 
-        arguments = self._generate_command_arguments(command)
-        command.arg_groups = self._group_arguments(arguments)
+        command.generate_args()
 
         return command
 
@@ -209,8 +205,7 @@ class CommandGenerator:
             put_op
         ]
 
-        arguments = self._generate_command_arguments(command)
-        command.arg_groups = self._group_arguments(arguments)
+        command.generate_args()
 
         group_name = self.generate_command_group_name_by_resource(
             resource_path=resource.path, rp_name=resource.resource_provider.name)
@@ -384,93 +379,6 @@ class CommandGenerator:
         else:
             raise NotImplementedError()
         return command_name
-
-    def _generate_command_arguments(self, command):
-        arguments = {}
-        for op in command.operations:
-            for arg in op.generate_args():
-                if arg.var not in arguments:
-                    arguments[arg.var] = arg
-
-        dropped_args = set()
-        used_args = set()
-        for arg in arguments.values():
-            used_args.add(arg.var)
-            if arg.var in dropped_args or not arg.options:
-                continue
-            r_arg = None
-            for v in arguments.values():
-                if v.var in used_args or v.var in dropped_args or arg.var == v.var or not v.options:
-                    continue
-                if not set(arg.options).isdisjoint(v.options):
-                    r_arg = v
-                    break
-            if r_arg:
-                if self._can_replace_argument(r_arg, arg):
-                    arg.ref_schema.arg = r_arg.var
-                    dropped_args.add(arg.var)
-                elif self._can_replace_argument(arg, r_arg):
-                    r_arg.ref_schema.arg = arg.var
-                    dropped_args.add(r_arg.var)
-                else:
-                    # let developer handle duplicated options
-                    logger.warning(
-                        f"Duplicated Option Value: {set(arg.options).intersection(r_arg.options)} : "
-                        f"{arg.var} with {r_arg.var} : {command.operations[-1].operation_id}"
-                    )
-
-        return [arg for var, arg in arguments.items() if var not in dropped_args]
-
-    @staticmethod
-    def _can_replace_argument(arg, old_arg):
-        arg_prefix = arg.var.split('.')[0]
-        old_prefix = old_arg.var.split('.')[0]
-        if old_prefix in (BuildInVariants.Query, BuildInVariants.Header, BuildInVariants.Path):
-            # replace argument should only be in body
-            return False
-        if arg_prefix in (BuildInVariants.Query, BuildInVariants.Header):
-            # only support path argument to replace
-            return False
-
-        elif arg_prefix == BuildInVariants.Path:
-            # path argument
-            arg_schema_required = arg.ref_schema.required
-            arg_schema_name = arg.ref_schema.name
-            try:
-                arg.ref_schema.required = old_arg.ref_schema.required
-                if old_arg.ref_schema.name == "name" and "name" in arg.options:
-                    arg.ref_schema.name = "name"
-                diff = arg.ref_schema.diff(old_arg.ref_schema, level=CMDDiffLevelEnum.Structure)
-                if diff:
-                    return False
-                return True
-            finally:
-                arg.ref_schema.name = arg_schema_name
-                arg.ref_schema.required = arg_schema_required
-        else:
-            # body argument
-            diff = arg.ref_schema.diff(old_arg.ref_schema, level=CMDDiffLevelEnum.Structure)
-            if diff:
-                return False
-            return True
-
-    @staticmethod
-    def _group_arguments(arguments):
-        arg_groups = {}
-        for arg in arguments:
-            group_name = arg.group or ""
-            if group_name not in arg_groups:
-                arg_groups[group_name] = {}
-            if arg.var not in arg_groups[group_name]:
-                arg_groups[group_name][arg.var] = arg
-
-        groups = []
-        for group_name, args in arg_groups.items():
-            group = CMDArgGroup()
-            group.name = group_name
-            group.args = [arg for arg in args.values()]
-            groups.append(group)
-        return groups or None
 
     # For update
     @staticmethod
